add func (ndb *nodeDB) getNodeFromCache(hash []byte) (n *Node)
